Map Native Staking calculated fields on transactions queue/history
