+++ src/routes/transactions/entities/staking/native-staking-info.entity.ts
@@ -44,6 +44,24 @@ export class NativeStakingDepositTransactionInfo
   annualNrr: number;
 
   @ApiProperty()
+  value: string;
+
+  @ApiProperty()
+  numValidators: number;
+
+  @ApiProperty()
+  expectedAnnualReward: string;
+
+  @ApiProperty()
+  expectedMonthlyReward: string;
+
+  @ApiProperty()
+  expectedFiatAnnualReward: number;
+
+  @ApiProperty()
+  expectedFiatMonthlyReward: number;
+
+  @ApiProperty()
   tokenInfo: TokenInfo;
 
   constructor(args: {
@@ -54,6 +72,12 @@ export class NativeStakingDepositTransactionInfo
     fee: number;
     monthlyNrr: number;
     annualNrr: number;
+    value: string;
+    numValidators: number;
+    expectedAnnualReward: string;
+    expectedMonthlyReward: string;
+    expectedFiatAnnualReward: number;
+    expectedFiatMonthlyReward: number;
     tokenInfo: TokenInfo;
   }) {
     super(TransactionInfoType.NativeStakingDeposit, null, null);
@@ -64,6 +88,12 @@ export class NativeStakingDepositTransactionInfo
     this.fee = args.fee;
     this.monthlyNrr = args.monthlyNrr;
     this.annualNrr = args.annualNrr;
+    this.value = args.value;
+    this.numValidators = args.numValidators;
+    this.expectedAnnualReward = args.expectedAnnualReward;
+    this.expectedMonthlyReward = args.expectedMonthlyReward;
+    this.expectedFiatAnnualReward = args.expectedFiatAnnualReward;
+    this.expectedFiatMonthlyReward = args.expectedFiatMonthlyReward;
     this.tokenInfo = args.tokenInfo;
   }
 }
