+++ src/routes/transactions/transactions-view.service.ts
@@ -1,7 +1,4 @@
 import { IConfigurationService } from '@/config/configuration.service.interface';
-import { IBalancesRepository } from '@/domain/balances/balances.repository.interface';
-import { IChainsRepository } from '@/domain/chains/chains.repository.interface';
-import { getNumberString } from '@/domain/common/utils/utils';
 import { IDataDecodedRepository } from '@/domain/data-decoder/data-decoded.repository.interface';
 import { DataDecoded } from '@/domain/data-decoder/entities/data-decoded.entity';
 import { ComposableCowDecoder } from '@/domain/swaps/contracts/decoders/composable-cow-decoder.helper';
@@ -27,7 +24,6 @@ import { Inject, Injectable } from '@nestjs/common';
 
 @Injectable({})
 export class TransactionsViewService {
-  private static readonly ETH_ETHERS_PER_VALIDATOR = 32;
   private readonly isNativeStakingEnabled: boolean;
 
   constructor(
@@ -45,10 +41,6 @@ export class TransactionsViewService {
     private readonly configurationService: IConfigurationService,
     private readonly kilnNativeStakingHelper: KilnNativeStakingHelper,
     private readonly nativeStakingMapper: NativeStakingMapper,
-    @Inject(IChainsRepository)
-    private readonly chainsRepository: IChainsRepository,
-    @Inject(IBalancesRepository)
-    private readonly balancesRepository: IBalancesRepository,
   ) {
     this.isNativeStakingEnabled = this.configurationService.getOrThrow<boolean>(
       'features.nativeStaking',
@@ -119,7 +111,7 @@ export class TransactionsViewService {
           ...nativeStakingTransaction,
           chainId: args.chainId,
           dataDecoded,
-          value: args.transactionDataDto.value,
+          value: args.transactionDataDto.value ?? null,
         });
       } else {
         // Should not reach here
@@ -286,40 +278,18 @@ export class TransactionsViewService {
     to: `0x${string}`;
     data: `0x${string}`;
     dataDecoded: DataDecoded;
-    value?: string;
+    value: string | null;
   }): Promise<NativeStakingDepositConfirmationView> {
     const depositInfo = await this.nativeStakingMapper.mapDepositInfo({
       chainId: args.chainId,
       to: args.to,
+      value: args.value,
       isConfirmed: false,
       depositExecutionDate: null,
     });
-    const value = args.value ? Number(args.value) : 0;
-    const chain = await this.chainsRepository.getChain(args.chainId);
-    const numValidators = Math.floor(
-      value /
-        Math.pow(10, chain.nativeCurrency.decimals) /
-        TransactionsViewService.ETH_ETHERS_PER_VALIDATOR,
-    );
-    const nativeCoinPrice =
-      await this.balancesRepository.getNativeCoinPrice(chain);
-
-    const expectedAnnualReward = (depositInfo.annualNrr / 100) * value;
-    const expectedMonthlyReward = expectedAnnualReward / 12;
-    const expectedFiatAnnualReward =
-      (expectedAnnualReward * (nativeCoinPrice ?? 0)) /
-      Math.pow(10, chain.nativeCurrency.decimals);
-    const expectedFiatMonthlyReward = expectedFiatAnnualReward / 12;
-
     return new NativeStakingDepositConfirmationView({
       method: args.dataDecoded.method,
       parameters: args.dataDecoded.parameters,
-      value: getNumberString(value),
-      numValidators,
-      expectedAnnualReward: getNumberString(expectedAnnualReward),
-      expectedMonthlyReward: getNumberString(expectedMonthlyReward),
-      expectedFiatAnnualReward,
-      expectedFiatMonthlyReward,
       ...depositInfo,
     });
   }
