+++ src/routes/transactions/transactions.module.ts
@@ -1,3 +1,4 @@
+import { BalancesRepositoryModule } from '@/domain/balances/balances.repository.interface';
 import { ChainsRepositoryModule } from '@/domain/chains/chains.repository.interface';
 import { ContractsRepositoryModule } from '@/domain/contracts/contracts.repository.interface';
 import { DataDecodedRepositoryModule } from '@/domain/data-decoder/data-decoded.repository.interface';
@@ -52,6 +53,7 @@ import { Module } from '@nestjs/common';
   controllers: [TransactionsController],
   imports: [
     AddressInfoModule,
+    BalancesRepositoryModule,
     ChainsRepositoryModule,
     ContractsRepositoryModule,
     DataDecodedRepositoryModule,
